git-remote-hg: add debug mode, activate via DEBUG_REMOTEHG env var
For now, all debug mode does is to let exceptions propagate to the top
(and thus print their backtrace), while in non-debug mode, backtraces
are now suppressed. This should be a bit more user friendly.

Additionally, we get rid of one try/except pair which replaced the
potentially helpful exception message with a bland 'Repository error';
now the user can e.g. distinguish whether a 404 error occurred,
authentication failed or the domain name is unknown (due to a typo).