Add Felipe's git-marks-check to the repository, for convenience
This doesn't solve the actual issues, but helps mitigate them in many
cases.

I am not completely sure what the "actual" issues are, but one seems to
be that it is apparently relatively easy to end up with a corrupt marks
file (and indeed, the code loading / saving marks does not attempt to be
particular robust).

The other problem hinted at in issue #4 seems to  be when tips in the
remote hg repository change, then this seems to also cause troubles.

See also:
  https://github.com/felipec/git-remote-hg/issues/4
  https://github.com/felipec/git-remote-hg/issues/10